refactored into online and offline game service
